move init function to ai.rs
